// types.ts
export interface Node {
  task_id: string;
  x: number;
  y: number;
  label: string;
}

export interface Example {
  id: string;
  primary_task_id: string;
  title: string;
  description: string;
  industry: string;
  complexity: 'Low' | 'Medium' | 'High';
  tags: string[];
  image_url: string;
  nodes: Node[];
}